requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
google-genai
jinja2